requirements.txt
flask==3.0.0
requests==2.31.0
grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.25.1
aiohttp==3.9.1
uvicorn==0.24.0
gunicorn==21.2.0
